Numbers: hook up watches and experiment with clamp
The intention of the clamping here is to specify clamping bounds against
the number itself, rather than needing to pass it to all the functions.
These clamp values are themselves our number type so they can
dynamically modified. As such, can use these for our measure type.

A bunch of tests & refactors to be made still.