Fix first-tick tooltip flicker
When recalculating tooltips when data changes, the first tick would take
the position based on the previous data's rendered desired size. This
fix hides the tooltip in the tick it is updated and forces a prepass so
that the subsequent tick calculates it position accurately, at which
point it is shown.