Fix border slot misuse
UBorderSlot assumes its Parent is a UBorder widget, which VulMultiBorder
is not. This caused crashes when altering the border slot's padding from
UMG editor. Fix is using a standard panel slot which has no UPROPERTY
controls, but this is fine for now as the multi border style does
support padding.