Support tooltip priority
For when multiple elements are competing for a tooltip render. This
allows these elements to declare a priority so that we can decide which
tooltip to show. This solution doesn't feel solid, in that we're needing
to force context clears. Suspect there are issues with this solution,
but will go with it for now.