Testing tool to preview icons quickly in editor
Note refactor to avoid circular dependency issues between VulRuntime and
VulEditor modules. This error presented as an error dialog on boot of
the editor project when it got to loading modules; complained about not
finding DDLs (that were there, but depended on each other). The fix for
now is to quickly move some code and have VulRuntime not require
VulEditor, though this might suggest a better code structure should be
considered in the future.