Level manager will now reload the current level
Previously a request to load the same level would hang on the loading
screen. Instead, now, we queue up a separate unload then load request.
This does mean that we cannot stream in and out concurrently like we do
for different level requests, but we have to wait for the level
streaming to actually get rid of the old level if we want a fresh one.