Spawn widgets in the player screen
If added to the viewport without a player, does not z-order correctly
with player-added widgets. This assumes games are adding their own
widgets to the player screen, and we want to be compatible with them.

Maybe this could be configurable in the future, but right now this works
with the single project using the library.