Hexgrid alternate trace support
When finding traces of tiles across hexgrids, allow some leeway to find
a non-blocking path if the trace line is close enough to it.

Particularly relevant for when the trace travels precisely along the
boundary between two hexes and either of the hexes would make a
non-blocked trace, but the other wouldn't.