Added start level override
So that a dev can change the starting level in the project settings,
rather than the vul level manager actor itself; which sits in the root
level and will cause conflicts when multiple devs work on it.

This allows a dev to work in their own level and have config point at
it for game start. Really it's probably better that our level manager
isn't added as an actor and we just configure a single-instance level
manager from project settings, but this is a quick workaround for now.