Level management is now a subsystem
This replaces the actor approach, which required the game to be played
from the root level. The motivation with this change is that the game
can be started from other levels in isolation (useful for development)
and still have core level manager-initiated functionality trigger, such
as spawning widgets.

With this change, it is easier for multiple developers to collaborate on
a game, keeping their editing in separate levels entirely.