Restoring temp values does not fail now
Previously the thinking was that we want ensure to indicate logic
errors, but in practice the inability to restore a temp value just means
that it was never stored (e.g. a new widget), so there's nothing to
restore it to. We could have introduced an optional return here, but
then the calling code would be cumbersome, undermining one of the
motivations for this template class in the first place.