Enhance TVulNumber modification capabilities
Now modifications can be overwritten by providing a modification with
the same ID, and it reports if this override actually occurred.

Modification IDs can now be customized to use something other than
FGuid, which is useful if you're applying well-known modifications in
your project that you want to refer to with a known ID (such as a
hardcoded FString).