Quick access to TVulNumber breakdown
This is useful when wanting to breakdown how a number is composed. E.g.
in a damage calculation context, an ability will have its base damage +
extra damage from strength then magic damage modifiers. This function
allows the caller to iterate over a simple delta for each modification
with the ID identifying what's the source of the modification.