Standardization of field APIs
This ensure we have consistent function signatures across fields, field
sets and type serializers.

Importantly this include a new de/serialization context, which both
provide the ability to track errors. These context structs are empty
right now, but are natural places for extension in the future without
needing to change signatures of everything.