Do not serialize data that is missing (by default).
This can be configured against entries in a field set. It was decided
that it only makes sense for this to exist in the context of a field set
(rather than field), as you need the containing object to decide whether
to include a property. For example, it doesn't make sense to return
invalid when serializing an empty string for a single string field.

Refactored the field set internal API so that it can offer a chained API
when defining fields. This is extenisble for further options in the
future.