Basic field descriptions & JSON schema
So we can generate schemas & tooling around our CPP/Unreal serialization
output without having to redefine everything.

This is just a start and demonstrates functionality with basic common
types (scalars, map, arrays), as well as integration with VulFieldSets
which can infer this information without consuming code needing to
update their definitions - as long as their types are
default-constructable; a requirement of deserialization already.

Included a simple JSON schema test to verify the implementation so far.