Schema generation now supports type hierarchies
Using a FVulFieldRegistry to statically define types & their
relationships, with some macros to do this with a little boilerplate as
possible. Only one abstract class + its concrete subtypes are supported
for now, generating schemas that have the base type as a OneOf all its
possible concrete types.

These concepts can likely be used to simplify deserialization, removing
the need for users to define their own deserialization functions for
types of this disciminator-based hierarchy.

This also adds support for recursive types, where an type contains a
field referencing itself. This is demonstrated with test coverage of our
polymorphic tree structure.

This logic was complicated to get working, and likely this can be
cleaned up significantly.