Extract referencing serialization mode
All objects that are referenced are extracted in to a dedicated root
node of serialized output. Updated serialization and meta output to
support this. Deserialization is not included, and the implementation is
a bit wonky: this is handled purely in the typescript & JSON schema
generation in the interest of time.