LevelManager: start of network support
This adds a bunch of utility logging, and gets to the point that PIE
clients get controlled by level manager. This is a bit tricky, as PIE
clients looking to spin on a new world up repeatedly until that can
connect to the real thing, it seems. The retry logic introduced here
solves this such that the clients eventually match the replicated world
Root world with all streaming data properly loaded.