Level manager: network-centric actor spawn system
This is not yet working as desired. Looking to get clients to have the
server spawn actors in for them, but cannot find a way to make the
client aware of them post-replication. Working to use a required actor
component that users will have to attach to their actors as a compromise
after multiple failed other approaches.